requirements.txt
modal
requests
ipykernel
feedparser
tqdm
bs4
pydantic
openai
dotenv
huggingface_hub 
numpy
torch
sentence-transformers 
chromadb                                                                                    
datasets
matplotlib
xgboost
scikit-learn==1.6.1
plotly
joblib
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
google-generativeai


